BUG: Fix incorrect prototype label usage
In AFLOW.solve_for_params_of_known_prototype, get_equivalent_atom_sets_from_prototype_and_atom_map was being called with the nominal prototype label, not the re-detected one.
This only matter in edge cases such as when the nominal label is A2B11_cP39_200_f_aghij. The equivalent label with the properly chosen minimal enumeration is A2B11_cP39_200_f_begik,
and the letters e,h form a non-alphabetically-contiguous Wyckoff set. This change in prototype label switches the position of this equivalent position in the string, breaking things
if the incorrect label is passed.